# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Treinamento API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/luan/games": {
            "get": {
                "tags": [
                    "Luan Games"
                ],
                "summary": "List all LuanGames",
                "operationId": "a26f62e25389aaa969578b131a1d7911",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Luan Games"
                ],
                "summary": "Create a new LuanGame",
                "operationId": "58502840f11da078280ac50525c4ed3f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "titulo": {
                                        "type": "string"
                                    },
                                    "genero_id": {
                                        "type": "string"
                                    },
                                    "plataforma_id": {
                                        "type": "string"
                                    },
                                    "ano_lancamento": {
                                        "type": "string"
                                    },
                                    "capa": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                }
            }
        },
        "/api/luan/games/{id}": {
            "get": {
                "tags": [
                    "Luan Games"
                ],
                "summary": "Get LuanGame by ID",
                "operationId": "857235912bdd362b3b4714a29aa74003",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Luan Games"
                ],
                "summary": "Update a LuanGame",
                "operationId": "8d130eb4da7a07a9be93cacb600cdbf3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "titulo": {
                                        "type": "string"
                                    },
                                    "genero_id": {
                                        "type": "string"
                                    },
                                    "plataforma_id": {
                                        "type": "string"
                                    },
                                    "ano_lancamento": {
                                        "type": "string"
                                    },
                                    "capa": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated successfully"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Luan Games"
                ],
                "summary": "Delete a LuanGame",
                "operationId": "69a2bcc0c1cf841cf6c52704c1e39aef",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted successfully"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/api/luan/genres": {
            "get": {
                "tags": [
                    "Luan Genres"
                ],
                "summary": "List all LuanGenres",
                "operationId": "913ec818ff9fae3176146b8ff0d2f4b5",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Luan Genres"
                ],
                "summary": "Create a new LuanGenre",
                "operationId": "b4846eea7a14beba895952bd84a0c25e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "genero": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                }
            }
        },
        "/api/luan/genres/{id}": {
            "get": {
                "tags": [
                    "Luan Genres"
                ],
                "summary": "Get LuanGenre by ID",
                "operationId": "94ab0b4aaa4db0a323770eaaaadcaaca",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Luan Genres"
                ],
                "summary": "Update a LuanGenre",
                "operationId": "b08e404ac14495e114dc29e7df90df98",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "genero": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated successfully"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Luan Genres"
                ],
                "summary": "Delete a LuanGenre",
                "operationId": "b921131bd8fb5b113c71e73f41016367",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted successfully"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
...
    },
...
}